requirement.txt
google-generativeai
reportlab
python-dotenv